Added Etterem class
The main frame now uses an Etterem object instead of etteremId and etteremNev
Asztal no longer stores information about Etterem, thus it has a new constructor without the etteremId, the etteremNev is now only a parameter used to set the asztalNev
Added EtteremValasztoDialog which is the first page to load. It loads all restaurants from the DB into a combobox and asks the user to select one. When the Ok button is pressed it sets the main frames etterem property to the selected Etterem